Refactor: Organize code into reusable components
